(Not related to DSL) - move IntegrationManagementConfigurer from management to config  - tangle between core and management
